Build fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@73646 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/AudioMixer.cpp
@@ -1760,7 +1760,7 @@ int audioMixerSinkWriteToStreams(PAUDMIXSINK pSink, AUDMIXOP enmOp, const void *
             Assert(cbToWriteStream >= cbChunk);
             cbToWriteStream -= (uint32_t)cbChunk;
 
-            cbWrittenStream += cbChunk;
+            cbWrittenStream += (uint32_t)cbChunk;
             Assert(cbWrittenStream <= cbBuf);
         }
